refactor: move trigger
